refactor: split backend startup and env wiring
