+++ docker-compose.prod.yml
@@ -27,7 +27,7 @@ services:
     container_name: votedots-backend
     restart: unless-stopped
     env_file:
-      - ./backend/.env
+      - ./backend/.env.server
     expose:
       - "4000"
     depends_on:
