+++ docker-compose.yml
@@ -8,7 +8,7 @@ services:
     ports:
       - "4000:4000"
     env_file:
-      - ./backend/.env
+      - ./backend/.env.docker
     depends_on:
       - redis
     volumes:
